requirements-linux.txt
aiohttp==3.9.3
aiosignal==1.3.1
albumentations==1.4.3
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.16.0
channels==4.1.0
colorama==0.4.6
constantly==23.10.4
contourpy==1.2.1
cryptography==42.0.6
cycler==0.12.1
daphne==4.1.2
debugpy==1.6.7
decorator==5.1.1
Django==5.0.4
djangorestframework==3.15.1
exceptiongroup==1.1.2
executing==1.2.0
filelock==3.13.4
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
grpcio
hyperlink==21.0.0
idna==3.6
imageio==2.34.0
importlib_metadata==6.8.0
incremental==22.10.0
ipykernel==6.25.2
ipython==8.14.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.0
jupyter_client==8.3.1
jupyter_core==5.3.1
kiwisolver==1.4.5
lazy_loader==0.4
lightning-utilities==0.11.2
Markdown==3.5
MarkupSafe==2.1.3
matplotlib==3.8.4
matplotlib-inline==0.1.6
mpmath==1.3.0
multidict==6.0.5
nest_asyncio==1.5.8
networkx==3.3
numpy==1.26.4
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==23.2
parso==0.8.3
pickleshare==0.7.5
pillow==10.3.0
platformdirs==3.11.0
polarTransform==2.0.0
prompt-toolkit==3.0.39
protobuf==3.20.1
psutil==5.9.5
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
Pygments==2.16.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
python-dateutil==2.8.2
pytorch-lightning==2.2.1
PyYAML==6.0.1
pyzmq==25.1.0
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.13.0
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.2
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.0
threadpoolctl==3.4.0
tifffile==2024.2.12
torch==2.2.2
torchaudio==2.2.2
torchmetrics==1.3.2
torchvision==0.17.2
tornado==6.3.3
tqdm==4.66.2
traitlets==5.14.3
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions
tzdata==2024.1
vtk==9.3.0
wcwidth==0.2.6
Werkzeug==3.0.0
yarl==1.9.4
zipp==3.17.0
zope.interface==6.3
